requirements.txt
# Google Gemini API
google-genai>=0.2.0

# Machine Learning (for difficulty prediction models)
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0

# Data handling
python-dateutil>=2.8.0

# MongoDB Database
pymongo>=4.6.0
python-dotenv>=1.0.0
